+++ arrow/src/array/array_fixed_size_binary.rs
